Mappings concluidos e ajustes nos modelos

+++ Muleki.Domain/Entities/PlayerFootball.cs
@@ -6,9 +6,9 @@ namespace Muleki.Domain.Entities
 {
     public class PlayerFootball : BaseEntity
     {
-        public int PlayerId { get; set; }
+        public long PlayerId { get; set; }
         public Player Player { get; set; }
-        public int FootballId { get; set; }
+        public long FootballId { get; set; }
         public Football Football { get; set; }
 
         // Entity Framework Core
