+++ Muleki.Domain/Entities/Safebox.cs
@@ -10,7 +10,7 @@ namespace Muleki.Domain.Entities
         public decimal Expense { get; set; }
         
         #region OneToOne
-        public int FootballId { get; set; }
+        public long FootballId { get; set; }
         public Football Football { get; set; }
         #endregion
